+++ modules/Food_Service/Users/ActivityReport.php
@@ -258,7 +258,7 @@ if ( ! $_REQUEST['modfunc'] )
 		'Transaction Types',
 		false,
 		$types_group,
-		[ 'save' => false, 'search' => false, 'print' => false ]
+		[ 'search' => false ]
 	);
 
 	ListOutput(
@@ -268,6 +268,6 @@ if ( ! $_REQUEST['modfunc'] )
 		'Transactions',
 		$link,
 		$group,
-		[ 'save' => false, 'search' => false, 'print' => false ]
+		[ 'search' => false ]
 	);
 }